Remove kamstrup special handling by moving the protocols to dedicated directories
