+++ conpot/protocols/kamstrup_management/kamstrup_management.xsd
