+++ conpot/protocols/kamstrup_management/kamstrup_management_server.py
@@ -19,9 +19,7 @@ import socket
 import gevent
 from gevent.server import StreamServer
 import conpot.core as conpot_core
-from conpot.protocols.kamstrup.management_protocol.command_responder import (
-    CommandResponder,
-)
+from .command_responder import CommandResponder
 from conpot.helpers import str_to_bytes
 from conpot.core.protocol_wrapper import conpot_protocol
