+++ conpot/protocols/kamstrup_meter/decoder_382.py
