+++ conpot/protocols/kamstrup_meter/kamstrup_constants.py
