+++ conpot/protocols/kamstrup_meter/kamstrup_meter.xsd
