+++ conpot/protocols/kamstrup_meter/messages.py
