+++ conpot/protocols/kamstrup_meter/register.py
