+++ conpot/protocols/kamstrup_meter/request_parser.py
