+++ conpot/templates/proxy/proxy/proxy.xml
@@ -2,12 +2,12 @@
     <proxy name="Kamstrup Channel A" host="0.0.0.0" port="1025">
         <proxy_host>192.168.1.176</proxy_host>
         <proxy_port>1025</proxy_port>
-        <decoder>conpot.protocols.kamstrup.meter_protocol.decoder_382.Decoder382</decoder>
+        <decoder>conpot.protocols.kamstrup_meter.decoder_382.Decoder382</decoder>
     </proxy>
     <proxy name="Kamstrup Channel B" host="0.0.0.0" port="1026">
         <proxy_host>192.168.0.10</proxy_host>
         <proxy_port>1026</proxy_port>
-        <decoder>conpot.protocols.kamstrup.meter_protocol.decoder_382.Decoder382</decoder>
+        <decoder>conpot.protocols.kamstrup_meter.decoder_382.Decoder382</decoder>
     </proxy>
     <proxy name="Kamstrup telnet" host="0.0.0.0" port="50100">
         <proxy_host>192.168.0.10</proxy_host>
