+++ conpot/tests/test_kamstrup_decoder.py
@@ -16,7 +16,7 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 import unittest
-from conpot.protocols.kamstrup.meter_protocol.decoder_382 import Decoder382
+from conpot.protocols.kamstrup_meter.decoder_382 import Decoder382
 
 
 class TestKamstrupDecoder(unittest.TestCase):
