+++ conpot/tests/test_kamstrup_meter.py
@@ -19,9 +19,9 @@ from gevent import monkey
 
 monkey.patch_all()
 import conpot
-from conpot.protocols.kamstrup.meter_protocol.command_responder import CommandResponder
-from conpot.protocols.kamstrup.meter_protocol import request_parser
-from conpot.protocols.kamstrup.meter_protocol.kamstrup_server import KamstrupServer
+from conpot.protocols.kamstrup_meter.command_responder import CommandResponder
+from conpot.protocols.kamstrup_meter.request_parser import KamstrupRequestParser
+from conpot.protocols.kamstrup_meter.kamstrup_server import KamstrupServer
 from conpot.helpers import chr_py3
 from conpot.utils.greenlet import spawn_test_server, teardown_test_server
 from gevent import socket
@@ -33,7 +33,7 @@ class TestKamstrup(unittest.TestCase):
     def setUp(self):
         # get the conpot directory
         self.dir_name = os.path.dirname(conpot.__file__)
-        self.request_parser = request_parser.KamstrupRequestParser()
+        self.request_parser = KamstrupRequestParser()
         self.command_responder = CommandResponder(
             self.dir_name + "/templates/kamstrup_382/kamstrup_meter/kamstrup_meter.xml"
         )
