--- docs/source/api/reference/conpot.protocols.kamstrup.management_protocol.rst
@@ -1,38 +1,0 @@
-conpot.protocols.kamstrup.management\_protocol package
-======================================================
-
-Submodules
-----------
-
-conpot.protocols.kamstrup.management\_protocol.command\_responder module
-------------------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.management_protocol.command_responder
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.management\_protocol.commands module
---------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.management_protocol.commands
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.management\_protocol.kamstrup\_management\_server module
-----------------------------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.management_protocol.kamstrup_management_server
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: conpot.protocols.kamstrup.management_protocol
-   :members:
-   :undoc-members:
-   :show-inheritance:
