--- docs/source/api/reference/conpot.protocols.kamstrup.meter_protocol.rst
@@ -1,70 +1,0 @@
-conpot.protocols.kamstrup.meter\_protocol package
-=================================================
-
-Submodules
-----------
-
-conpot.protocols.kamstrup.meter\_protocol.command\_responder module
--------------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol.command_responder
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.meter\_protocol.decoder\_382 module
--------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol.decoder_382
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.meter\_protocol.kamstrup\_constants module
---------------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol.kamstrup_constants
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.meter\_protocol.kamstrup\_server module
------------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol.kamstrup_server
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.meter\_protocol.messages module
----------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol.messages
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.meter\_protocol.register module
----------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol.register
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-conpot.protocols.kamstrup.meter\_protocol.request\_parser module
-----------------------------------------------------------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol.request_parser
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: conpot.protocols.kamstrup.meter_protocol
-   :members:
-   :undoc-members:
-   :show-inheritance:
