--- docs/source/api/reference/conpot.protocols.kamstrup.rst
@@ -1,21 +1,0 @@
-conpot.protocols.kamstrup package
-=================================
-
-Subpackages
------------
-
-.. toctree::
-
-   conpot.protocols.kamstrup.management_protocol
-   conpot.protocols.kamstrup.meter_protocol
-
-Submodules
-----------
-
-Module contents
----------------
-
-.. automodule:: conpot.protocols.kamstrup
-   :members:
-   :undoc-members:
-   :show-inheritance:
