+++ docs/source/api/reference/conpot.protocols.kamstrup_management.rst
@@ -1,0 +1,38 @@
+conpot.protocols.kamstrup\_management package
+=============================================
+
+Submodules
+----------
+
+conpot.protocols.kamstrup\_management.command\_responder module
+---------------------------------------------------------------
+
+.. automodule:: conpot.protocols.kamstrup_management.command_responder
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+conpot.protocols.kamstrup\_management.commands module
+-----------------------------------------------------
+
+.. automodule:: conpot.protocols.kamstrup_management.commands
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+conpot.protocols.kamstrup\_management.kamstrup\_management\_server module
+-------------------------------------------------------------------------
+
+.. automodule:: conpot.protocols.kamstrup_management.kamstrup_management_server
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: conpot.protocols.kamstrup_management
+   :members:
+   :undoc-members:
+   :show-inheritance:
